+++ lib/chat/file_fs.ex
@@ -92,11 +92,21 @@ defmodule Chat.FileFs do
   def relative_filenames(prefix) do
     dir = build_path(prefix)
 
-    dir
-    |> list_files()
-    |> Enum.flat_map(&populate_level/1)
-    |> Enum.flat_map(&populate_level/1)
-    |> Enum.map(&String.slice(&1, (String.length(dir) + 1)..-1))
+    if File.dir?(dir) do
+      dir
+      |> list_files()
+      |> Enum.flat_map(&populate_level/1)
+      |> Enum.flat_map(&populate_level/1)
+      |> Enum.map(&String.slice(&1, (String.length(dir) + 1)..-1))
+    else
+      []
+    end
+  end
+
+  def has_file?({_, _, _} = keys, prefix \\ nil) do
+    keys
+    |> file_path(build_path(prefix))
+    |> File.exists?()
   end
 
   ##
@@ -107,6 +117,8 @@ defmodule Chat.FileFs do
     path
     |> File.ls!()
     |> Enum.map(&Path.join([path, &1]))
+  rescue
+    _ -> []
   end
 
   defp list_files(path) do