+++ lib/chat_web/live/main_live/page/admin_panel_router.ex
@@ -28,6 +28,9 @@ defmodule ChatWeb.MainLive.Page.AdminPanelRouter do
 
       {"unmount-main", _} ->
         socket |> AdminPanel.unmount_main()
+
+      {"toggle-gpio24-impendance", _} ->
+        socket |> AdminPanel.toggle_gpio24_impendance()
     end
   end
 
@@ -53,6 +56,21 @@ defmodule ChatWeb.MainLive.Page.AdminPanelRouter do
 
       {:unmounted_main, _} ->
         socket
+
+      :refresh_rooms_and_users ->
+        socket |> AdminPanel.refresh_rooms_and_users()
+
+      {:gpio24_impedance_status, value} ->
+        socket |> AdminPanel.set_gpio24_impedance_status(value)
+
+      {:connect_to_weight_sensor, name, opts} ->
+        socket |> AdminPanel.connect_to_weight_sensor(name, opts)
+
+      {:weight_sensor_connection, status} ->
+        socket |> AdminPanel.weight_sensor_connection_status(status)
+
+      {:create_cargo_user, name} ->
+        socket |> AdminPanel.create_cargo_user(name)
     end
   end